refactor: FillList -> FillListWithRepeatOperation 이름 변경
