+++ src/main/java/lotto/function/FillListWithRepeatOperation.java
@@ -5,7 +5,7 @@ import java.util.List;
 import java.util.function.BiFunction;
 import java.util.function.Supplier;
 
-public class FillList<T> implements BiFunction<Integer, Supplier<T>, List<T>> {
+public class FillListWithRepeatOperation<T> implements BiFunction<Integer, Supplier<T>, List<T>> {
 
 	@Override
 	public List<T> apply(Integer index, Supplier<T> supplier) {
